staging: android: ion: Remove check of debug_file
There's no need to check the return value of debug_file for it is just a
debugfs and we will go on the following process if we failed to create
debug_file. So just remove it.

Signed-off-by: Yisheng Xie <xieyisheng1@huawei.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Panchajanya Sarkar <panchajanya@azure-dev.live>
Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>
(cherry picked from commit 0540dfb13792ed1e8a3bdae6431a09e6a0bfe3b2)
Signed-off-by: rezaadi0105 <rezaadipangestu5@gmail.com>